+++ clients/admin-ui/src/pages/consent/reporting/index.tsx
@@ -1,9 +1,7 @@
-import { Box, Breadcrumb, BreadcrumbItem, Heading, Text } from "@fidesui/react";
-import NextLink from "next/link";
+import { Box, Heading, Text } from "@fidesui/react";
 import React from "react";
 
 import Layout from "~/features/common/Layout";
-import { CONSENT_REPORTING_ROUTE } from "~/features/common/nav/v2/routes";
 import ConsentReporting from "~/features/consent-reporting/ConsentReporting";
 
 const ConsentReportingPage = () => (
@@ -12,21 +10,6 @@ const ConsentReportingPage = () => (
       <Heading fontSize="2xl" fontWeight="semibold" mb={2} data-testid="header">
         Configure consent
       </Heading>
-      <Box>
-        <Breadcrumb
-          fontWeight="medium"
-          fontSize="sm"
-          color="gray.600"
-          data-testid="breadcrumbs"
-        >
-          <BreadcrumbItem>
-            <NextLink href={CONSENT_REPORTING_ROUTE}>Consent</NextLink>
-          </BreadcrumbItem>
-          <BreadcrumbItem color="complimentary.500">
-            <NextLink href="#">Reporting</NextLink>
-          </BreadcrumbItem>
-        </Breadcrumb>
-      </Box>
     </Box>
     <Text fontSize="sm" mb={8} width={{ base: "100%", lg: "50%" }}>
       Download a CSV containing a report of consent preferences made by users on